requirements.txt
beautifulsoup4
pillow
requests
tenacity
matplotlib
discord-py-interactions
